disable pivkey on container build
