+++ cmd/cosign/cli/pivcli/commands.go
@@ -1,3 +1,5 @@
+// +build !pivkeydisabled
+
 // Copyright 2021 The Sigstore Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
