+++ pkg/cosign/keys.go
@@ -30,7 +30,6 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/go-piv/piv-go/piv"
 	"github.com/pkg/errors"
 	"github.com/theupdateframework/go-tuf/encrypted"
 
@@ -144,38 +143,6 @@ func LoadECDSAPrivateKey(key []byte, pass []byte) (signature.ECDSASignerVerifier
 		return signature.ECDSASignerVerifier{}, fmt.Errorf("invalid private key")
 	}
 	return signature.NewECDSASignerVerifier(epk, crypto.SHA256), nil
-}
-func GetYubikey() (*piv.YubiKey, error) {
-	cards, err := piv.Cards()
-	if err != nil {
-		return nil, err
-	}
-
-	// Find a YubiKey and open the reader.
-	var yk *piv.YubiKey
-	for _, card := range cards {
-		if strings.Contains(strings.ToLower(card), "yubikey") {
-			if yk, err = piv.Open(card); err != nil {
-				return nil, err
-			}
-			return yk, nil
-		}
-	}
-	return nil, errors.New("no yubikey found")
-}
-
-func GenYubikey(yk *piv.YubiKey) (crypto.PublicKey, error) {
-	// Generate a private key on the YubiKey.
-	key := piv.Key{
-		Algorithm:   piv.AlgorithmEC256,
-		PINPolicy:   piv.PINPolicyAlways,
-		TouchPolicy: piv.TouchPolicyAlways,
-	}
-	pub, err := yk.GenerateKey(piv.DefaultManagementKey, piv.SlotSignature, key)
-	if err != nil {
-		return nil, err
-	}
-	return pub, nil
 }
 
 const pubKeyPemType = "PUBLIC KEY"
